refactor: Back test api 수정
