+++ src/main/java/com/traders/traders/module/history/controller/HistoryController.java
@@ -1,9 +1,15 @@
 package com.traders.traders.module.history.controller;
 
+import com.traders.traders.module.history.controller.dto.request.BackTestRequestDto;
+import com.traders.traders.module.history.controller.dto.response.BackTestResponseDto;
 import com.traders.traders.module.history.service.HistoryService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import javax.validation.Valid;
 
 @RestController
 @RequiredArgsConstructor
@@ -11,5 +17,8 @@ import org.springframework.web.bind.annotation.RestController;
 public class HistoryController {
     private final HistoryService historyService;
 
-    //TODO - 다른 전략이랑 한 화면에서 비교를 해줄 수도 있고!!
+    @GetMapping("")
+    public BackTestResponseDto backTest(@Valid @ModelAttribute BackTestRequestDto request) {
+        return historyService.backTest(request.toServiceDto());
+    }
 }
