+++ src/main/java/com/traders/traders/module/strategy/service/StrategyService.java
@@ -3,29 +3,24 @@ package com.traders.traders.module.strategy.service;
 import com.traders.traders.common.exception.TradersException;
 import com.traders.traders.common.utils.AESUtils;
 import com.traders.traders.module.feign.service.FeignService;
-import com.traders.traders.module.history.domain.repository.dao.HistoryDao;
 import com.traders.traders.module.history.service.HistoryService;
 import com.traders.traders.module.strategy.controller.dto.request.CreateStrategyDto;
-import com.traders.traders.module.strategy.controller.dto.response.BackTestResponseDto;
 import com.traders.traders.module.strategy.controller.dto.response.FindStrategiesInfoResponseDto;
 import com.traders.traders.module.strategy.domain.Position;
 import com.traders.traders.module.strategy.domain.Strategy;
 import com.traders.traders.module.strategy.domain.repository.StrategyRepository;
 import com.traders.traders.module.strategy.domain.repository.dao.StrategyInfoDao;
-import com.traders.traders.module.strategy.service.dto.*;
+import com.traders.traders.module.strategy.service.dto.WebHookDto;
 import com.traders.traders.module.users.domain.Users;
 import com.traders.traders.module.users.domain.repository.dao.AutoTradingSubscriberDao;
 import com.traders.traders.module.users.service.UsersService;
 import com.traders.traders.module.users.service.dto.SubscribeStrategyDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
 import static com.traders.traders.common.exception.ExceptionMessage.NOT_FOUND_ANY_STRATEGY_EXCEPTION;
@@ -41,7 +36,6 @@ public class StrategyService {
     private final UsersService userService;
     private final StrategyRepository strategyRepository;
     private final AESUtils aesUtils;
-    private final RedisTemplate<String, Object> redisTemplate;
 
     @Async //TODO - 동기 비동기 성능 차이 블로그 작성하기
     public void handleWebHook(WebHookDto request) {
@@ -80,29 +74,6 @@ public class StrategyService {
         savedUser.subscribeStrategy(strategy, encryptedApiKey, encryptedSecretKey);
     }
 
-    public BackTestResponseDto backTest(BackTestDto request) {
-
-        List<StrategyHistoryDto> strategyHistoryDtos = new ArrayList<>();
-
-        for (Long id : request.getIds()) {
-            String cacheKey = "strategyId:" + id;
-
-            //캐시에 매매 내역들 있는지 확인
-            HistoryCache historyCache = (HistoryCache) redisTemplate.opsForValue().get(cacheKey);
-
-            //없다면 매매 내역 전체 캐싱
-            if (historyCache == null) {
-                List<HistoryDao> histories = findHistoryDaoByStrategyId(id);
-                historyCache = HistoryCache.of(histories);
-                redisTemplate.opsForValue().set(cacheKey, historyCache);
-            }
-
-            //매매 내역들을 조건에 맞게 (기간, 자기 자본 100) 계산 후 응답
-            strategyHistoryDtos.add(calculateHistoryCache(historyCache, request, id));
-        }
-
-        return BackTestResponseDto.of(strategyHistoryDtos);
-    }
 
     public void createStrategy(CreateStrategyDto request) {
         Strategy strategy = Strategy.of(request.getName(), request.getProfitFactor(), request.getWinningRate(),
@@ -165,45 +136,5 @@ public class StrategyService {
     private List<StrategyInfoDao> findStrategyInfoDaos() {
         return strategyRepository.findStrategiesInfoDao()
                 .orElseThrow(() -> new TradersException(NOT_FOUND_ANY_STRATEGY_EXCEPTION));
-    }
-
-    private List<HistoryDao> findHistoryDaoByStrategyId(Long id) {
-        return historyService.findHistoryDaoByStrategyId(id);
-    }
-
-    private StrategyHistoryDto calculateHistoryCache(HistoryCache historyCache, BackTestDto request,
-                                                     Long strategyId) {
-        Long id = strategyId;
-        double compoundProfitRate = 0;
-        double winCount = 0;
-        double totalCount = 0; //TODO - 기간을 이상하게 설정해서 매매 내역이 없는 경우 0으로 나누게 됐을 때 예외 발생시키기
-        double totalProfitRate = 0;
-        double totalLossRate = 0;
-        //TODO - 숏 총 순익, 롱 총 순익 추가하기
-
-        for (HistoryDao history : historyCache.getHistories()) {
-            if (isInPeriod(history, request.getStartDate(), request.getEndDate())) {
-                totalCount++;
-                compoundProfitRate = compoundProfitRate * (1 + history.getProfitRate());
-
-                if (history.getProfitRate() > 0) {
-                    winCount++;
-                    totalProfitRate += history.getProfitRate();
-                } else {
-                    totalLossRate += history.getProfitRate();
-                }
-            }
-        }
-
-        StrategyInfoDto strategyInfoDto = StrategyInfoDto.of(id, compoundProfitRate, winCount / totalCount,
-                totalProfitRate / totalLossRate);
-
-        return StrategyHistoryDto.of(strategyInfoDto, historyCache.getHistories());
-
-    }
-
-    private boolean isInPeriod(HistoryDao history, LocalDateTime startDate, LocalDateTime endDate) {
-        return history.getEntryPosition().getTime().isAfter(startDate) &&
-                history.getExitPosition().getTime().isBefore(endDate);
     }
 }
