--- src/main/java/com/traders/traders/module/strategy/service/dto/StrategyHistoryDto.java
@@ -1,19 +1,0 @@
-package com.traders.traders.module.strategy.service.dto;
-
-import java.util.List;
-
-import com.traders.traders.module.history.domain.repository.dao.HistoryDao;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class StrategyHistoryDto {
-	private final StrategyInfoDto strategyInfoDto;
-	private final List<HistoryDao> historyDaos;
-
-	public static StrategyHistoryDto of(StrategyInfoDto strategyInfoDto, List<HistoryDao> historyDaos) {
-		return new StrategyHistoryDto(strategyInfoDto, historyDaos);
-	}
-}
